Set css styles on page instead of per element
Since the kingdom map is larger and takes longer to update, the js method of adding per element inline styles kicked off before the tiles were recreated. This resulted in having to add a delay to make sure the style adding logic ran after the tiles rendered. Since we don't know when the rendering finished, we had to guess.

This new method simply creates a set of styles to be added in the head <style> tag and lets the browser apply the styles automatically. No need to rerun the style highlighter when ever user interacts with the UI.